+++ core/security.py
@@ -1,0 +1,12 @@
+# 비밀번호 해싱을 위한 컨텍스트 객체
+from passlib.context import CryptContext
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+# 비밀번호 암호화
+def get_password_hash(password: str) -> str:
+    return pwd_context.hash(password)
+
+# 비밀번호 검증
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    return pwd_context.verify(plain_password, hashed_password) # 텍스트 비번-해시 비번간 일치 여부 반환
